Newly Updated
Corrected a few more issues with my button features. Working on making headers and links same size, color, and location as the original webpage. Also, all of the buttons and links are linked to the actual sites links and buttons. Still need to work on the side by side image and the footer. Closing in.